Add form for insert massages

+++ bboard/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Bb(models.Model):
     title = models.CharField(max_length=50, verbose_name='Товар')
     content = models.TextField(null=True, blank=True, verbose_name='Описание')
-    price = models.FloatField(null=True, blank=True, verbose_name='Цена')
+    price = models.DecimalField(null=True, blank=True, verbose_name='Цена', max_digits=12, decimal_places=2)
     published = models.DateTimeField(auto_now_add=True, db_index=True, verbose_name='Опубликовано')
     rubric = models.ForeignKey('Rubric', null=True, on_delete=models.PROTECT, verbose_name='Рубрика')
